Split out the cors stuff and move pre-flight into the application controller
