+++ app/controllers/Application.scala
@@ -10,6 +10,7 @@ import play.api.data.Forms._
 // Pan domain
 import com.gu.pandomainauth.action.AuthActions
 import com.gu.pandomainauth.model.AuthenticatedUser
+import helpers.CORSable
 
 trait PanDomainAuthActions extends AuthActions {
 
@@ -31,6 +32,8 @@ trait PanDomainAuthActions extends AuthActions {
 
 object Application extends Controller with PanDomainAuthActions {
 
+  lazy val composer = config.getString("composer.domain").get
+
   val urlForm = Form(
     "url" -> nonEmptyText
   )
@@ -50,4 +53,13 @@ object Application extends Controller with PanDomainAuthActions {
     )
   }
 
+  def preflight(routes: String) = CORSable(composer) {
+    Action { implicit req =>
+      val requestedHeaders = req.headers("Access-Control-Request-Headers")
+
+      NoContent.withHeaders(
+        CORSable.CORS_ALLOW_METHODS -> "GET, DELETE, PUT",
+        CORSable.CORS_ALLOW_HEADERS -> requestedHeaders)
+    }
+  }
 }
