+++ app/helpers/CORSable.scala
@@ -1,0 +1,30 @@
+package helpers
+
+import play.api.mvc._
+import play.api.libs.json._
+import scala.concurrent.Future
+import scala.concurrent.ExecutionContext.Implicits.global
+
+/* Must allow cross origin requests for the templating service */
+case class CORSable[A](origins: String*)(action: Action[A]) extends Action[A] {
+  def apply(request: Request[A]): Future[Result] = {
+
+    val headers = request.headers.get("Origin").map { origin =>
+      if(origins.contains(origin)) {
+        List("Access-Control-Allow-Origin" -> origin, "Access-Control-Allow-Credentials" -> "true")
+      } else { Nil }
+    }
+
+    action(request).map(_.withHeaders(headers.getOrElse(Nil) :_*))
+  }
+
+  lazy val parser = action.parser
+}
+
+
+object CORSable {
+  val CORS_ALLOW_ORIGIN = "Access-Control-Allow-Origin"
+  val CORS_CREDENTIALS = "Access-Control-Allow-Credentials"
+  val CORS_ALLOW_METHODS = "Access-Control-Allow-Methods"
+  val CORS_ALLOW_HEADERS = "Access-Control-Allow-Headers"
+}
